NEW: Prepare new code structure for sharing code
Separate the high-level code from the low-level code so
that the high-level code can be shared between the simulation
and the real microcontroller.

The plan is to make this git repo a submodule of the simulator
repo and use a define BUILD_MCU to distinguish between simulator
build and microcontroller build.

Also fix some code styling - my taste has changed since the
last time I touched this code.