ENH: Reduce ROM footprint, UART tracing and improve ADC
Many things in a single commit.

First of, reduce ROM footprint by removing usage of stdlib printf
and double.

Replace printf with a trace function, which uses the printf lib by
mpaland under the hood for the MCU (and regular vprintf for simulation).
Configure the printf lib to make it as small as possible. Also add a
UART driver so that we can print the traces to the console via the
Launchpad UART->USB, baud 9600 Hz.

Accidently used double on a few places, replace with float. In the
future we should use fixed point instead because float is slow and
also takes up a lot of space.

Secondly, rework the ADC handling. Instead of doing the sampling and
conversion when we call the function, use interrupt instead and
write the results to a ring buffer of size 3 for each channel. We don't
have to sample it that often because the sensors themselves are slow
(~40 ms), so configure it to use the ACLK (~8-32 kHz). There sudden
spikes in the values, so use a median filter to remove those. This
makes it much better, but it might not be necessary once we add a
capacitor. We must also sample channel 5 and 6 instead of 1 and 2
now because those pins are occupied by UART.

There is a shortage of timers on the MSP430G2553, and since we don't
use the watchdog timer for reset, reconfigure it for regular interrupts
instead to support a time elapsed function (time_ms).

Also
* Extend makefile with useful commands (size and readelf)
* Add functions for driving the LEDs on the launchpad
* Add useful defines UNUSED and ARRAY_SIZE
* Fix GPIO selection (it was incomplete so couldn't select all modes)