FIX: GPIO idx, PWM period and remote control state machine
I accidently used the wrong value for indexing GPIO pins.
There is a couple of bits left in the 16-bit used for storing
GPIO information, use them to store the index. Create a new
macro/define to retrieve the index.

Integrate ir remote control in the new state machine. Handle
the commands in a separate sub state machine.

After getting the remote control to work again I noticed the
motors are running very slow. The reason is that they
don't get the expected voltage, which is caused by a too low
PWM frequency. It's actually very low (100 Hz). It works as
it should when increasing it to 10 kHz.

Add two defines for two levels of tracing (warn/info).