FIX: State machine working with line sensors
Some small fixes to make the state machine code work on the
real sumobot. It can now drive around on the platform avoiding
the white border.

Also, create a test to confirm that the new range sensors work.
Rename the enemy_detection to opponent_detection because of some
weird ccstudio bug causing it to not pick up the symbols.