ENH: Refactor search state and fix ADC bug
Refactor SEARCH_1 and SEARCH_2 states into a single main state
SEARCH and have it call a sub-state machine like RETREAT.

ADC may stop working after a while. The reason is unclear, but
after a while the interrupts stops coming. The interrupt is what
triggers the next sampling/conversion so if we miss one interrupt
it would stop working, but it's unclear why we would miss the
a single interrupt. Disabling global interrupts instead of just
the ADC interrupt seems to fix it, so do that.

Disable enemy detection for now.