ENH: vl53l0x: Multi-read using interrupt
The next step in reducing the time we spend reading the
vl53l0x sensors is to use the interrupt from the front
sensor. The approach is to start range measuring all sensors
(starting with the front sensor) and then set a bool from
the interrupt when front sensor is done, and only then
start read the measurements from all sensors. Thus, most
of the time we call vl53l0x_read... we return the previous
values. This prevents the line detection from being blocked
by the range sensors.

Also, always set both selection bits when changing the pin
function, otherwise we may end up with an incorrect selection
if we reconfigure.