NEW: state_machine: Detect history and shared state data
Make Nsumo remember where the enemy and line was detected
and use it determine where to drive. This improves enemy
following and make it easier to "breakout" of bad situations.

Move all data saved acroess loop iterations to a single
struct (state_machine_data_t). Collect the input data once at
the beginning and then only pass the data the main states need.
Allocate this data on the stack so that the main function
can be re-used by several simulated sumobots (mcus).