ENH: Better arc turning and fix 16 MHz IR remote
Introduce three types of arcturns, sharp, mid, wide, and have
them mapped to three different speeds just like rotation and
forward. Clean up the code that maps drive+speed to duty cycles.

IR remote also depends on SMCLK since it uses a timer, so it's
broken after SMCLK was changed to 16 MHz. Adjust the values of
the timer to make it work again. While at it, clean up the todos
and sharpen the code.