FIX: Refactor state machine into multiple files
The state_machine.c contains a lot of code, which makes it
harder to read. Break the different states into separate files
and put the common code in a common header and only keep the
main state handling in state_machine.c.

Also clean up the includes between simulator and MCU. Add
target includes to simulator so that we can use the same
include path when they include the same file.